python/requirements.txt
SQLAlchemy>=1.4.40
PyYAML>=6.0
Faker>=13.15.0
numpy>=1.21.0
simpy>=4.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
psycopg2-binary>=2.9.5
flask>=2.0.0
flask-cors>=3.0.0 